Add Trips#show skeleton and play around with Backbone
Previously, the page on /trips/new was listing items. Instead, we'll
want to have separate pages for creating a trip/items and for displaying
a trip.

The show page is still a skeleton - the DB is not being queried to find
the items associated with a trip, although the URL parameter is in
place which will enable that query.

Backbone.js is added as an experiment - it may or may not be necessary,
but its event handling is pretty neat and may fit in well with the
client-heavy nature of the app.